# data/swagger.json
{"openapi":"3.0.0","paths":{"/exam-api/register":{"post":{"operationId":"RegistrantsController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrantCreateDto"}}}},"responses":{"201":{"description":"Successfully registered"}},"summary":"Register a new candidate","tags":["exam"]}},"/exam-api/{registrantId}":{"put":{"operationId":"RegistrantsController_updateRegistrant","parameters":[{"name":"registrantId","required":true,"in":"path","description":"Registrant ID","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PartialTypeClass"}}}},"responses":{"200":{"description":""}},"summary":"Update registrant details","tags":["exam"]},"delete":{"operationId":"RegistrantsController_removeRegistrant","parameters":[{"name":"registrantId","required":true,"in":"path","description":"Registrant ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Delete a registrant","tags":["exam"]}},"/exam-api":{"get":{"operationId":"RegistrantsController_allRegistrants","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentGradeLevel","required":false,"in":"query","schema":{"type":"string"}},{"name":"cadre","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examStatus","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examType","required":false,"in":"query","schema":{"type":"string"}},{"name":"expectedRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"gender","required":false,"in":"query","schema":{"type":"string"}},{"name":"mda","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get all registrants with pagination","tags":["exam"]}},"/exam-api/registrant/{registrantId}":{"get":{"operationId":"RegistrantsController_registrantDetails","parameters":[{"name":"registrantId","required":true,"in":"path","description":"Registrant ID","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get registrant details by ID","tags":["exam"]}},"/exam-api/dashboard":{"get":{"operationId":"RegistrantsController_dashboard","parameters":[],"responses":{"200":{"description":""}},"summary":"Get dashboard statistics","tags":["exam"]}},"/exam-api/promotion-stats":{"get":{"operationId":"RegistrantsController_promotionStats","parameters":[],"responses":{"200":{"description":""}},"summary":"Get promotion statistics","tags":["exam"]}},"/exam-api/levels-status":{"get":{"operationId":"RegistrantsController_examStatusByLevel","parameters":[],"responses":{"200":{"description":""}},"summary":"Get promotion statistics","tags":["exam"]}},"/exam-api/status/passed":{"get":{"operationId":"RegistrantsController_getPasses","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentGradeLevel","required":false,"in":"query","schema":{"type":"string"}},{"name":"cadre","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examStatus","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examType","required":false,"in":"query","schema":{"type":"string"}},{"name":"expectedRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"gender","required":false,"in":"query","schema":{"type":"string"}},{"name":"mda","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get passed registrants","tags":["exam"]}},"/exam-api/status/failed":{"get":{"operationId":"RegistrantsController_getFailures","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentGradeLevel","required":false,"in":"query","schema":{"type":"string"}},{"name":"cadre","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examStatus","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examType","required":false,"in":"query","schema":{"type":"string"}},{"name":"expectedRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"gender","required":false,"in":"query","schema":{"type":"string"}},{"name":"mda","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get failed registrants","tags":["exam"]}},"/exam-api/status/incapacitated":{"get":{"operationId":"RegistrantsController_getDisabled","parameters":[{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentGradeLevel","required":false,"in":"query","schema":{"type":"string"}},{"name":"cadre","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examStatus","required":false,"in":"query","schema":{"type":"string"}},{"name":"exam.examType","required":false,"in":"query","schema":{"type":"string"}},{"name":"expectedRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"presentRank","required":false,"in":"query","schema":{"type":"string"}},{"name":"gender","required":false,"in":"query","schema":{"type":"string"}},{"name":"mda","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"Get registrants with disabilities","tags":["exam"]}},"/exam-api/auth/signup":{"post":{"operationId":"UserController_signup","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"fullname":{"type":"string"},"department":{"type":"string"},"jobTitle":{"type":"string"},"workEmailAddress":{"type":"string"},"password":{"type":"string"},"profilePicture":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/exam-api/auth/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}}},"info":{"title":"Exam API","description":"The Exam API documentation","version":"1.0","contact":{}},"tags":[],"servers":[{"url":"/"}],"components":{"schemas":{"QualificationDto":{"type":"object","properties":{"qualification":{"type":"string","example":"Bachelor of Science","description":"The qualification title"},"dateOfQualification":{"format":"date-time","type":"string","example":"2020-01-01","description":"The date when the qualification was obtained"}},"required":["qualification","dateOfQualification"]},"ExamCreateDto":{"type":"object","properties":{"examDate":{"format":"date-time","type":"string","example":"2024-02-07","description":"The date of the examination"},"examType":{"type":"string","example":"promotion","description":"Type of examination","enum":["promotion","conversion","confirmation"]},"examStatus":{"type":"string","example":"passed","description":"Status of the examination","enum":["passed","failed"]},"generalPaperScore":{"type":"number","example":75,"description":"Score obtained in general paper"},"professionalPaperScore":{"type":"number","example":80,"description":"Score obtained in professional paper"},"remark":{"type":"string","example":"Excellent performance","description":"Remarks about the examination performance"},"examNumber":{"type":"string","example":"EX2024001","description":"Unique examination number"}},"required":["examDate","examType","examStatus","generalPaperScore","professionalPaperScore","remark","examNumber"]},"RegistrantCreateDto":{"type":"object","properties":{"surname":{"type":"string","example":"Doe","description":"The surname of the registrant"},"firstName":{"type":"string","example":"John","description":"The first name of the registrant"},"middleName":{"type":"string","example":"James","description":"The middle name of the registrant"},"dateOfBirth":{"format":"date-time","type":"string","example":"1990-01-01","description":"The date of birth of the registrant"},"staffVerificationNumber":{"type":"string","example":"SVN123456","description":"The staff verification number of the registrant"},"gender":{"type":"string","example":"Male","description":"Gender of the registrant"},"presentRank":{"type":"string","example":"Sergeant","description":"The current rank of the registrant"},"expectedRank":{"type":"string","example":"Inspector","description":"The expected rank of the registrant"},"presentGradeLevel":{"type":"string","example":"GL-8","description":"The present grade level of the registrant"},"presentStep":{"type":"string","example":"5","description":"The present step of the registrant"},"expectedGradeLevel":{"type":"string","example":"GL-9","description":"The expected grade level of the registrant"},"dateOfPrevAppointment":{"format":"date-time","type":"string","example":"2020-01-01","description":"The date of the previous appointment of the registrant"},"dateOfConfirmation":{"format":"date-time","type":"string","example":"2018-01-01","description":"The date of confirmation of the registrant"},"dateOfPresentAppointment":{"format":"date-time","type":"string","example":"2020-01-01","description":"The date of the present appointment of the registrant"},"dateOfFirstAppointment":{"format":"date-time","type":"string","example":"2015-01-01","description":"The date of the first appointment of the registrant"},"disability":{"type":"boolean","example":false,"description":"The disability status of the registrant"},"qualifications":{"description":"The qualifications of the registrant","type":"array","items":{"$ref":"#/components/schemas/QualificationDto"}},"profilePassport":{"type":"string","example":"https://example.com/profile.jpg","description":"The profile passport of the registrant"},"email":{"type":"string","example":"john.doe@example.com","description":"The email address of the registrant"},"phone":{"type":"string","example":"08012345678","description":"The phone number of the registrant"},"cadre":{"type":"string","example":"Inspector","description":"The cadre of the registrant"},"mda":{"type":"string","example":"Ministry of Finance","description":"The MDA of the registrant"},"exam":{"description":"The examination details of the registrant","allOf":[{"$ref":"#/components/schemas/ExamCreateDto"}]}},"required":["surname","firstName","gender","presentRank","disability","email","phone","cadre","mda"]},"PartialTypeClass":{"type":"object","properties":{"surname":{"type":"string","example":"Doe","description":"The surname of the registrant"},"firstName":{"type":"string","example":"John","description":"The first name of the registrant"},"middleName":{"type":"string","example":"James","description":"The middle name of the registrant"},"dateOfBirth":{"format":"date-time","type":"string","example":"1990-01-01","description":"The date of birth of the registrant"},"staffVerificationNumber":{"type":"string","example":"SVN123456","description":"The staff verification number of the registrant"},"gender":{"type":"string","example":"Male","description":"Gender of the registrant"},"presentRank":{"type":"string","example":"Sergeant","description":"The current rank of the registrant"},"expectedRank":{"type":"string","example":"Inspector","description":"The expected rank of the registrant"},"presentGradeLevel":{"type":"string","example":"GL-8","description":"The present grade level of the registrant"},"presentStep":{"type":"string","example":"5","description":"The present step of the registrant"},"expectedGradeLevel":{"type":"string","example":"GL-9","description":"The expected grade level of the registrant"},"dateOfPrevAppointment":{"format":"date-time","type":"string","example":"2020-01-01","description":"The date of the previous appointment of the registrant"},"dateOfConfirmation":{"format":"date-time","type":"string","example":"2018-01-01","description":"The date of confirmation of the registrant"},"dateOfPresentAppointment":{"format":"date-time","type":"string","example":"2020-01-01","description":"The date of the present appointment of the registrant"},"dateOfFirstAppointment":{"format":"date-time","type":"string","example":"2015-01-01","description":"The date of the first appointment of the registrant"},"disability":{"type":"boolean","example":false,"description":"The disability status of the registrant"},"qualifications":{"description":"The qualifications of the registrant","type":"array","items":{"$ref":"#/components/schemas/QualificationDto"}},"profilePassport":{"type":"string","example":"https://example.com/profile.jpg","description":"The profile passport of the registrant"},"email":{"type":"string","example":"john.doe@example.com","description":"The email address of the registrant"},"phone":{"type":"string","example":"08012345678","description":"The phone number of the registrant"},"cadre":{"type":"string","example":"Inspector","description":"The cadre of the registrant"},"mda":{"type":"string","example":"Ministry of Finance","description":"The MDA of the registrant"},"exam":{"description":"The examination details of the registrant","allOf":[{"$ref":"#/components/schemas/ExamCreateDto"}]}}},"UserLoginDto":{"type":"object","properties":{"email":{"type":"string","example":"Carlo.Romaguera0@yahoo.com"},"password":{"type":"string","description":"string password","example":"hgevp7RQDF@@!123"}},"required":["email","password"]}}}}